+++ src/main/kotlin/net/ccbluex/liquidbounce/config/types/group/ValueGroup.kt
@@ -592,57 +592,6 @@ open class ValueGroup(
         fun ValueGroup.build()
     }
 
-    fun modes(
-        name: String,
-        default: String,
-        modes: Map<String, ModeBuilder>,
-    ): ModeValueGroup<Mode> {
-        require(modes.isNotEmpty()) { "Mode group '$name' must contain at least one mode." }
-        class SimpleMode(name: String, override val parent: ModeValueGroup<*>) : Mode(name)
-
-        return modes(
-            eventListener = null,
-            name = name,
-            activeCallback = { modes ->
-                val idx = modes.indexOfFirst { it.name == default }
-
-                check(idx != -1) {
-                    "The active choice $default is not contained within the choice array (${modes.joinToString { it.name }})"
-                }
-
-                idx
-            },
-            modesCallback = { parent ->
-                modes.entries.mapToArray { (modeName, configure) ->
-                    val mode = SimpleMode(modeName, parent)
-
-                    with(configure) {
-                        mode.build()
-                    }
-                    mode
-                }
-            },
-        )
-    }
-
-    private fun jsonChoice(name: String, valueObject: JsonObject) {
-        val value = valueObject["value"].asString
-        val modes = valueObject["choices"].asJsonArray.associateTo(linkedMapOf()) { choiceElement ->
-            val choiceObject = choiceElement.asJsonObject
-            val choiceName = choiceObject["name"].asString
-            val settings = choiceObject["values"]?.asJsonArray ?: emptyList<JsonElement>()
-            val configure = ModeBuilder {
-                for (setting in settings) {
-                    json(setting.asJsonObject)
-                }
-            }
-
-            choiceName to configure
-        }
-
-        modes(name, value, modes)
-    }
-
     protected fun <T : Mode> modes(
         eventListener: EventListener?,
         name: String,
@@ -653,14 +602,15 @@ open class ValueGroup(
             val idx = modes.indexOf(active)
 
             check(idx != -1) {
-                "The active choice $active is not contained within the choice array (${modes.joinToString { it.name }})"
+                "The active choice $active is not contained within the choice array" +
+                    " (${modes.joinToString { it.name }})"
             }
 
             idx
         }) { modes }
     }
 
-    protected fun <T : Mode> modes(
+    fun <T : Mode> modes(
         eventListener: EventListener?,
         name: String,
         activeCallback: ToIntFunction<List<T>>,
@@ -680,180 +630,5 @@ open class ValueGroup(
     ) = modes(eventListener, name, { activeIndex }, choicesCallback)
 
     fun <V : Value<*>> value(value: V) = value.apply { this@ValueGroup.inner.add(this) }
-
-    /**
-     * Assigns the value of the settings to the component
-     *
-     * A component can have dynamic settings which can be assigned through the JSON file
-     * These have to be interpreted and assigned to the value group
-     *
-     * An example:
-     * {
-     *     "type": "INT",
-     *     "name": "Size",
-     *     "value": 14,
-     *     "range": {
-     *         "min": 1,
-     *         "max": 100
-     *     },
-     *     "suffix": "px"
-     * }
-     *
-     * TODO: Replace with proper deserialization
-     *
-     * @param valueObject JsonObject
-     */
-    @Suppress("LongMethod")
-    fun json(valueObject: JsonObject) {
-        val type = enumValueOf<ValueType>(valueObject["type"].asString)
-        val name = valueObject["name"].asString
-
-        // todo: replace this with serious deserialization
-        when (type) {
-            ValueType.BOOLEAN -> {
-                val value = valueObject["value"].asBoolean
-                boolean(name, value)
-            }
-
-            ValueType.INT -> {
-                val value = valueObject["value"].asInt
-                val min = valueObject["range"].asJsonObject["min"].asInt
-                val max = valueObject["range"].asJsonObject["max"].asInt
-                val suffix = valueObject["suffix"]?.asString ?: ""
-                int(name, value, min..max, suffix)
-            }
-
-            ValueType.INT_RANGE -> {
-                val valueMin = valueObject["value"].asJsonObject["min"].asInt
-                val valueMax = valueObject["value"].asJsonObject["max"].asInt
-                val min = valueObject["range"].asJsonObject["min"].asInt
-                val max = valueObject["range"].asJsonObject["max"].asInt
-                val suffix = valueObject["suffix"]?.asString ?: ""
-                intRange(name, valueMin..valueMax, min..max, suffix)
-            }
-
-            ValueType.FLOAT -> {
-                val value = valueObject["value"].asFloat
-                val min = valueObject["range"].asJsonObject["min"].asFloat
-                val max = valueObject["range"].asJsonObject["max"].asFloat
-                val suffix = valueObject["suffix"]?.asString ?: ""
-                float(name, value, min..max, suffix)
-            }
-
-            ValueType.FLOAT_RANGE -> {
-                val valueMin = valueObject["value"].asJsonObject["min"].asFloat
-                val valueMax = valueObject["value"].asJsonObject["max"].asFloat
-                val min = valueObject["range"].asJsonObject["min"].asFloat
-                val max = valueObject["range"].asJsonObject["max"].asFloat
-                val suffix = valueObject["suffix"]?.asString ?: ""
-                floatRange(name, valueMin..valueMax, min..max, suffix)
-            }
-
-            ValueType.TEXT -> {
-                val value = valueObject["value"].asString
-                text(name, value)
-            }
-
-            ValueType.FILE -> {
-                val value = valueObject["value"]
-                    ?.takeUnless { it is JsonNull }
-                    ?.asString?.takeUnless(String::isBlank)
-                    ?.let(::File)
-                val dialogMode = valueObject["dialogMode"]
-                    ?.takeUnless { it is JsonNull }
-                    ?.asString
-                    ?.let(FileDialogMode::valueOf)
-                    ?: FileDialogMode.OPEN_FILE
-                val supportedExtensions = valueObject["supportedExtensions"]
-                    ?.takeUnless { it is JsonNull }
-                    ?.asJsonArray
-                    ?.mapTo(linkedSetOf()) { it.asString }
-
-                file(name, value, dialogMode, supportedExtensions)
-            }
-
-            ValueType.COLOR -> {
-                val value = valueObject["value"].asInt
-                color(name, Color4b(value))
-            }
-
-            ValueType.CONFIGURABLE -> {
-                val subValueGroup = ValueGroup(name)
-                val values = valueObject["values"].asJsonArray
-                for (value in values) {
-                    subValueGroup.json(value.asJsonObject)
-                }
-                tree(subValueGroup)
-            }
-            // same as value group but it is [ToggleableValueGroup]
-            ValueType.TOGGLEABLE -> {
-                val value = valueObject["value"].asBoolean
-                // Parent is NULL in that case because we are not dealing with Listenable anyway and only use it
-                // as toggleable ValueGroup
-                val subValueGroup = object : ToggleableValueGroup(null, name, value) {}
-                val settings = valueObject["values"].asJsonArray
-                for (setting in settings) {
-                    subValueGroup.json(setting.asJsonObject)
-                }
-                tree(subValueGroup)
-            }
-
-            ValueType.CHOOSE -> {
-                val value = valueObject["value"].asString.asTagged()
-                val choices = valueObject["choices"].asJsonArray.mapTo(linkedSetOf()) { it.asString.asTagged() }
-
-                enumChoice(name, value, choices)
-            }
-
-            ValueType.CHOICE -> jsonChoice(name, valueObject)
-
-            ValueType.MULTI_CHOOSE -> {
-                fun parseBoolean(key: String, default: Boolean) = when (val json = valueObject[key]) {
-                    null, is JsonNull -> default
-                    is JsonPrimitive, is JsonArray -> json.asBoolean
-                    else -> error("Unexpected JSON value (${json.javaClass}): $json, should be boolean")
-                }
-
-                val canBeNone = parseBoolean(key = "canBeNone", default = true)
-                val isOrderSensitive = parseBoolean(key = "isOrderSensitive", default = false)
-
-                val value = valueObject["value"].asJsonArray.mapTo(
-                    if (isOrderSensitive) linkedSetOf() else sortedSetOf()
-                ) { it.asString.asTagged() }
-                val choices = valueObject["choices"].asJsonArray.mapTo(linkedSetOf()) { it.asString.asTagged() }
-
-                multiEnumChoice(name, default = value, choices, canBeNone, isOrderSensitive)
-            }
-
-            ValueType.REGISTRY_LIST -> {
-                val innerValueType = enumValueOf<ValueType>(valueObject["innerValueType"].asString)
-                val normalizedValue = when (val value = valueObject["value"]) {
-                    is JsonArray -> value
-                    is JsonPrimitive -> listOf(value)
-                    null, is JsonNull -> emptyList()
-                    else -> error("Unexpected JSON value (${value.javaClass}): $value, should be Identifier list")
-                }
-
-                when (innerValueType) {
-                    ValueType.BLOCK -> {
-                        blocks(name, normalizedValue.mapTo(blockSortedSetOf()) {
-                            publicGson.fromJson(it, Block::class.java)
-                        })
-                    }
-
-                    ValueType.ITEM -> {
-                        items(name, normalizedValue.mapTo(itemSortedSetOf()) {
-                            publicGson.fromJson(it, Item::class.java)
-                        })
-                    }
-
-                    else -> error("Unsupported inner value type for ${ValueType.REGISTRY_LIST}: $innerValueType")
-                }
-            }
-
-            else -> error("Unsupported type: $type")
-        }
-    }
-
 
 }
