+++ src/main/kotlin/net/ccbluex/liquidbounce/integration/theme/component/components/WebHudComponent.kt
@@ -20,6 +20,7 @@
 package net.ccbluex.liquidbounce.integration.theme.component.components
 
 import com.google.gson.JsonObject
+import net.ccbluex.liquidbounce.config.types.group.json
 import net.ccbluex.liquidbounce.integration.theme.component.HudComponent
 import net.ccbluex.liquidbounce.integration.theme.component.HudComponentTweak
 import net.ccbluex.liquidbounce.utils.render.Alignment
